requirement.txt
numpy
retrying
python-dotenv
openai
scikit-learn
